Fix `NilClass` method error
When using this gem with a site that does not define any custom
redirects in the configuration file, it tries to call the method
`each` on a `nil` it gets by looking up `custom_redirects` in the
config file.  Instead, we want to use the local variable we created
to check for this specific case, instead of looking it up again.